Fixed turn direction choice, by:
Calculating the distance traveled for both a left and a right first turn

I'm not sure why I was trying to figure out which one was right with all sorts of overly complicated math. It's easier just to check both and take the shorter path or perhapse even in the future, the path that requires the least deceleration while obeying radius constraints.